+++ libnormaliz/lineare_transformation.h
@@ -46,7 +46,7 @@ namespace libnormaliz {
 
 template<typename Integer>
 class Lineare_Transformation {
-  int rk;
+  size_t rk;
   string status;
   Integer index;
   Matrix<Integer> Center;
@@ -68,7 +68,7 @@ public:
 //---------------------------------------------------------------------------
 
   void read() const;                   // to be modified, just for tests
-  int get_rank() const;              //returns rank if status is
+  size_t get_rank() const;              //returns rank if status is
 							   // "initialized, after transformation"
   string get_status()const;       //returns status, may be:
 							 // "non initialized"
@@ -78,7 +78,7 @@ public:
   Matrix<Integer> get_center() const;          //read center matrix
   Matrix<Integer> get_right() const;          //read right matrix
   Matrix<Integer> get_right_inv() const;     //read the inverse of the right matrix
-  void set_rank(const int rank);
+  void set_rank(const size_t rank);
   void set_center(const Matrix<Integer>& M);          //write center matrix
   void set_right(const Matrix<Integer>& M);          //write right matrix
   void set_right_inv(const Matrix<Integer>& M);     //write the inverse of the right matrix
@@ -87,15 +87,15 @@ public:
 //					  Rows and columns exchange
 //---------------------------------------------------------------------------
 
-  void exchange_rows(int row1, int row2);     //similar to Matrix<Integer>::exchange_rows
-  void exchange_columns(int col1, int col2); //similar to Matrix<Integer>::exchange_columns
+  void exchange_rows(size_t row1, size_t row2);     //similar to Matrix<Integer>::exchange_rows
+  void exchange_columns(size_t col1, size_t col2); //similar to Matrix<Integer>::exchange_columns
 
 //---------------------------------------------------------------------------
 //					Rows and columns reduction
 //---------------------------------------------------------------------------
 
-  void reduce_row(int corner);      //similar to Matrix<Integer>::reduce_row
-  void reduce_column(int corner);  //similar to Matrix<Integer>::reduce_column
+  void reduce_row(size_t corner);      //similar to Matrix<Integer>::reduce_row
+  void reduce_column(size_t corner);  //similar to Matrix<Integer>::reduce_column
 
 //---------------------------------------------------------------------------
 //					 Algorithms
@@ -111,7 +111,7 @@ public:
   /* test the main computation for arithmetic overflow
    * uses multiplication mod m
    */
-  bool test_transformation(const Matrix<Integer>& M, const int& m) const;
+  bool test_transformation(const Matrix<Integer>& M, const size_t& m) const;
 
 
 };
